MAJOR CHANGE: Revisited Unicode and UTF-8 handling when parsing JSON.
Prior to that change, the JSON parser was very sensible to non-ascii
characters and was throwing exception when it encoutered a non-ascii
character. To allow UTF-8 characters, we had to use the Utf8String that
allowed the parser to decode Unicode codepoints (UTF-8 encoded). After
experiencing many issues with that class and UTF-8 decoding in general,
we decided to completly remove that class and assume that, by default,
any std::string could contain non-ascii characters.

This now means that any std::string can contain UTF-8 encoded characters
and special iterators have to be used to iterate over such a string.