AugmentationLoop now decrements the inFlight counter when expiring an entry.
Prior to this patch, inFlight was never decremented if an inflight
augmenting request timeout. This means that we could end up in a state
when the AugmentationLoop stopped sending requests to a particular
instance because it accumulated too much in flights that were never
decremented, even after expiring. This patch now fixes this issue.